feat: add review summary subcomponents
