+++ src/components/reviews/ReviewSummaryPillars.tsx
@@ -1,0 +1,44 @@
+import * as React from "react";
+import type { Pillar } from "@/lib/types";
+import SectionLabel from "@/components/reviews/SectionLabel";
+import PillarBadge from "@/components/ui/league/pillars/PillarBadge";
+
+function StaticNeonWrap({ children }: { children: React.ReactNode }) {
+  return (
+    <span className="relative inline-flex">
+      <span
+        aria-hidden
+        className="pointer-events-none absolute inset-0 rounded-2xl bg-[linear-gradient(90deg,hsl(var(--accent)),hsl(var(--primary)))] opacity-40 blur-[6px]"
+      />
+      <span
+        aria-hidden
+        className="pointer-events-none absolute inset-0 rounded-2xl ring-1 ring-[hsl(var(--ring)/.35)]"
+      />
+      <span className="relative z-10">{children}</span>
+    </span>
+  );
+}
+
+export type ReviewSummaryPillarsProps = {
+  pillars?: Pillar[];
+};
+
+export default function ReviewSummaryPillars({ pillars }: ReviewSummaryPillarsProps) {
+  return (
+    <div>
+      <SectionLabel>Pillars</SectionLabel>
+      {Array.isArray(pillars) && pillars.length > 0 ? (
+        <div className="flex flex-wrap gap-2">
+          {pillars.map((p) => (
+            <StaticNeonWrap key={p}>
+              <PillarBadge pillar={p} size="md" active />
+            </StaticNeonWrap>
+          ))}
+        </div>
+      ) : (
+        <div className="text-sm text-muted-foreground">No pillars selected.</div>
+      )}
+    </div>
+  );
+}
+
