+++ src/components/reviews/ReviewSummaryScore.tsx
@@ -1,0 +1,82 @@
+import * as React from "react";
+import { cn } from "@/lib/utils";
+import SectionLabel from "@/components/reviews/SectionLabel";
+import NeonIcon from "@/components/reviews/NeonIcon";
+
+export type ReviewSummaryScoreProps = {
+  score: number;
+  msg: string;
+  ScoreIcon: React.ComponentType<{ className?: string }>;
+  scoreIconCls?: string;
+  focusOn?: boolean;
+  focus?: number;
+  focusMsg?: string;
+};
+
+export default function ReviewSummaryScore({
+  score,
+  msg,
+  ScoreIcon,
+  scoreIconCls,
+  focusOn,
+  focus,
+  focusMsg,
+}: ReviewSummaryScoreProps) {
+  return (
+    <div>
+      <SectionLabel>Score</SectionLabel>
+      <div className="relative h-12 rounded-2xl border border-[hsl(var(--border))] bg-[hsl(var(--card))] px-4">
+        <div className="absolute left-4 right-4 top-1/2 -translate-y-1/2">
+          <div className="relative h-2 w-full rounded-full bg-[hsl(var(--muted))] shadow-[inset_2px_2px_4px_hsl(var(--shadow-color)/0.45),inset_-2px_-2px_4px_hsl(var(--foreground)/0.06)]">
+            <div
+              className="absolute left-0 top-0 h-2 rounded-full shadow-[0_0_8px_hsl(var(--primary)/0.5)]"
+              style={{
+                width: `calc(${score * 10}% + 10px)`,
+                background: "linear-gradient(90deg, hsl(var(--primary)), hsl(var(--accent)))",
+              }}
+            />
+            <div
+              className="absolute top-1/2 h-5 w-5 -translate-y-1/2 rounded-full border border-[hsl(var(--border))] bg-[hsl(var(--card))] shadow-[0_10px_25px_hsl(var(--shadow-color)/.25)]"
+              style={{ left: `calc(${score * 10}% - 10px)` }}
+            />
+          </div>
+        </div>
+      </div>
+      <div className="mt-1 flex items-center gap-2 text-[13px] text-muted-foreground">
+        <span className="pill h-6 px-2 text-xs">{score}/10</span>
+        <ScoreIcon className={cn("h-4 w-4", scoreIconCls)} />
+        <span>{msg}</span>
+      </div>
+      {focusOn && typeof focus === "number" && focusMsg && (
+        <div className="mt-4 space-y-2">
+          <div className="mb-2 flex items-center gap-2" aria-label="Focus">
+            <NeonIcon kind="brain" on={true} size={32} staticGlow />
+            <div className="h-px flex-1 bg-gradient-to-r from-[hsl(var(--foreground)/0.20)] via-[hsl(var(--foreground)/0.05)] to-transparent" />
+          </div>
+          <div className="relative h-12 rounded-2xl border border-[hsl(var(--border))] bg-[hsl(var(--card))] px-4">
+            <div className="absolute left-4 right-4 top-1/2 -translate-y-1/2">
+              <div className="relative h-2 w-full rounded-full bg-[hsl(var(--muted))] shadow-[inset_2px_2px_4px_hsl(var(--shadow-color)/0.45),inset_-2px_-2px_4px_hsl(var(--foreground)/0.06)]">
+                <div
+                  className="absolute left-0 top-0 h-2 rounded-full shadow-[0_0_8px_hsl(var(--accent)/0.5)]"
+                  style={{
+                    width: `calc(${(focus / 10) * 100}% + 10px)`,
+                    background: "linear-gradient(90deg, hsl(var(--accent)), hsl(var(--primary)))",
+                  }}
+                />
+                <div
+                  className="absolute top-1/2 h-5 w-5 -translate-y-1/2 rounded-full border border-[hsl(var(--border))] bg-[hsl(var(--card))] shadow-[0_10px_25px_hsl(var(--shadow-color)/.25)]"
+                  style={{ left: `calc(${(focus / 10) * 100}% - 10px)` }}
+                />
+              </div>
+            </div>
+          </div>
+          <div className="mt-1 flex items-center gap-2 text-[13px] text-muted-foreground">
+            <span className="pill h-6 px-2 text-xs">{focus}/10</span>
+            <span>{focusMsg}</span>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
+
