+++ src/components/reviews/index.ts
@@ -4,6 +4,11 @@ export { default as ReviewsPage } from "./ReviewsPage";
 export { default as ReviewEditor } from "./ReviewEditor";
 export { default as ReviewList } from "./ReviewList";
 export { default as ReviewSummary } from "./ReviewSummary";
+export { default as ReviewSummaryHeader } from "./ReviewSummaryHeader";
+export { default as ReviewSummaryScore } from "./ReviewSummaryScore";
+export { default as ReviewSummaryPillars } from "./ReviewSummaryPillars";
+export { default as ReviewSummaryTimestamps } from "./ReviewSummaryTimestamps";
+export { default as ReviewSummaryNotes } from "./ReviewSummaryNotes";
 export { default as RoleSelector } from "./RoleSelector";
 export { default as ReviewPanel } from "./ReviewPanel";
 export { default as NeonIcon } from "./NeonIcon";
